chore: rename sampler/ to player/
