+++ src/index.ts
@@ -1,8 +1,8 @@
 export * from "./drum-machine/drum-machine";
 export * from "./electric-piano";
 export * from "./mallet";
+export * from "./player/sampler";
 export * from "./reverb/reverb";
-export * from "./sampler/sampler";
 export * from "./soundfont/soundfont";
 export * from "./splendid-grand-piano";
 export * from "./storage";
