+++ src/player/signals.ts
